+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog for Telegram Bot Platform
 
+#### 1.2 / 2018-01-20
+
+* Added check for the presence of a method in the class, for the context.
+* Added checking of the parameters and their typing in the method, for the context.
+
 #### 1.1 / 2017-12-15
 
 * Added recognition `/cmd@botname` multiple bots are in a group.